# docs/api/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.availableModel:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
  admin.availableModelsResponse:
    properties:
      models:
        items:
          $ref: '#/definitions/admin.availableModel'
        type: array
    type: object
  admin.changeOwnPasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
...
  admin.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  admin.loginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/admin.meResponse'
    type: object
  admin.meResponse:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      is_system_admin:
        type: boolean
      role:
        type: string
    type: object
...
  admin.swaggerErrorResponse:
    properties:
      error:
        $ref: '#/definitions/apierror.SwaggerDetail'
    type: object
...
  apierror.SwaggerDetail:
    properties:
      code:
        example: not_found
        type: string
      message:
        example: resource not found
        type: string
    type: object
info:
  contact: {}
  description: Prepaid LLM API marketplace with OpenAI-compatible proxy, wallet billing,
    and model catalog.
  title: Tavo API
  version: 0.2.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies credentials and returns a 24-hour session token. Revokes
        any existing session for the user.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      summary: Authenticate with email and password
      tags:
      - auth
...
  /me:
    get:
      description: Returns the profile of the user associated with the current session
        key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.meResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get authenticated user profile
      tags:
      - auth
  /me/available-models:
    get:
      description: Returns models accessible to the authenticated key.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.availableModelsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List models available to the authenticated key
      tags:
      - auth
  /me/password:
    post:
      consumes:
      - application/json
      description: Verifies the current password and sets a new one. Revokes all other
        active sessions. Requires a user-scoped session key.
      parameters:
      - description: Password change request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.changeOwnPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/admin.swaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Change own password
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: API key (vl_uk_) or session key (vl_sk_)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...