# internal/api/admin/auth.go
package admin

import (
	"errors"
	"log/slog"
	"time"

	"github.com/gofiber/fiber/v3"
	"golang.org/x/crypto/bcrypt"

	"github.com/jukaza/tavo/internal/apierror"
	"github.com/jukaza/tavo/internal/audit"
	"github.com/jukaza/tavo/internal/auth"
	"github.com/jukaza/tavo/internal/db"
	"github.com/jukaza/tavo/internal/jsonx"
)

// marshalDescription serializes a map to compact JSON for audit event descriptions.
func marshalDescription(v any) string {
	b, err := jsonx.Marshal(v)
	if err != nil {
		return ""
	}
	return string(b)
}

// dummyHash is used to burn CPU time on failed login paths, preventing
// timing-based user enumeration (valid email ~100ms vs invalid email ~0ms).
var dummyHash, _ = bcrypt.GenerateFromPassword([]byte("void-dummy-timing-pad"), bcrypt.DefaultCost)

// auditLoginFailed records a failed login attempt in the audit log.
func (h *Handler) auditLoginFailed(c fiber.Ctx, email string, statusCode int) {
	if h.AuditLogger == nil {
		return
	}
	h.AuditLogger.Log(audit.Event{
		Timestamp:    time.Now().UTC(),
		Action:       "login_failed",
		ResourceType: "session",
		Description:  marshalDescription(map[string]string{"email": email}),
		IPAddress:    c.IP(),
		StatusCode:   statusCode,
		RequestID:    apierror.RequestIDFromCtx(c),
	})
}

// loginRequest is the JSON body for POST /api/v1/auth/login.
type loginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// meResponse is the JSON body returned for the authenticated user's profile.
type meResponse struct {
	ID                 string `json:"id"`
	Email              string `json:"email"`
	DisplayName        string `json:"display_name"`
	Role               string `json:"role"`
	IsSystemAdmin      bool   `json:"is_system_admin"`
	HasPassword        bool   `json:"has_password"`
	AuthProvider       string `json:"auth_provider"`
	ActiveSessionCount int    `json:"active_session_count"`
}

// loginResponse is the JSON body returned on successful authentication.
type loginResponse struct {
	Token     string     `json:"token"`
	ExpiresAt string     `json:"expires_at"`
	User      meResponse `json:"user"`
}

// Login handles POST /api/v1/auth/login. It verifies email and password and
// returns a short-lived session token valid for 24 hours. The session token is
// a real api_keys row and works with the existing auth middleware.
// This endpoint does not require prior authentication.
//
// @Summary      Authenticate with email and password
// @Description  Verifies credentials and returns a 24-hour session token. Revokes any existing session for the user.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        body  body      loginRequest  true  "Login credentials"
// @Success      200   {object}  loginResponse
// @Failure      400   {object}  swaggerErrorResponse
// @Failure      401   {object}  swaggerErrorResponse
// @Failure      500   {object}  swaggerErrorResponse
// @Router       /auth/login [post]
func (h *Handler) Login(c fiber.Ctx) error {
	var req loginRequest
	if err := c.Bind().JSON(&req); err != nil {
		h.auditLoginFailed(c, req.Email, fiber.StatusBadRequest)
		return apierror.BadRequest(c, "invalid request body")
	}
	if req.Email == "" {
		h.auditLoginFailed(c, req.Email, fiber.StatusBadRequest)
		return apierror.BadRequest(c, "email is required")
	}
	if req.Password == "" {
		h.auditLoginFailed(c, req.Email, fiber.StatusBadRequest)
		return apierror.BadRequest(c, "password is required")
	}

	// Brute-force protection: per-IP rate limit and per-account lockout.
	// Called after field validation so that missing-field 400s are not throttled.
	if h.LoginThrottle != nil {
		if err := h.LoginThrottle.Allow(c.IP(), req.Email); err != nil {
			h.Log.LogAttrs(c.Context(), slog.LevelWarn, "login throttled",
				slog.String("ip", c.IP()),
				slog.String("email", req.Email),
			)
			// Burn bcrypt time before returning 429 so that a throttled response
			// takes ~the same wall time as a normal failed-password response.
			// Without this, an attacker can distinguish "account locked" (fast)
			// from "wrong password" (slow ~100ms) via timing. The error is
			// intentionally discarded — the burn's only purpose is elapsed time.
			_ = bcrypt.CompareHashAndPassword(dummyHash, []byte(req.Password))
			return apierror.Send(c, fiber.StatusTooManyRequests, "too_many_requests", "too many login attempts, try again later")
		}
	}

	ctx := c.Context()

	userID, hash, err := h.DB.GetUserPasswordHash(ctx, req.Email)
	if err != nil {
		// Burn bcrypt time to prevent timing-based email enumeration.
		_ = bcrypt.CompareHashAndPassword(dummyHash, []byte(req.Password))
		if errors.Is(err, db.ErrNotFound) || errors.Is(err, db.ErrNoPassword) {
			if h.LoginThrottle != nil {
				h.LoginThrottle.RecordFailure(req.Email)
			}
			h.auditLoginFailed(c, req.Email, fiber.StatusUnauthorized)
			return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "invalid email or password")
		}
		h.Log.ErrorContext(ctx, "login: get user password hash", slog.String("error", err.Error()))
		return apierror.InternalError(c, "authentication failed")
	}

	if err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(req.Password)); err != nil {
		if h.LoginThrottle != nil {
			h.LoginThrottle.RecordFailure(req.Email)
		}
		h.auditLoginFailed(c, req.Email, fiber.StatusUnauthorized)
		return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "invalid email or password")
	}

	if h.LoginThrottle != nil {
		h.LoginThrottle.RecordSuccess(req.Email)
	}

	sess, err := h.issueUserSession(c, userID, "login")
	if err != nil {
		return apierror.InternalError(c, "authentication failed")
	}

	return c.Status(fiber.StatusOK).JSON(sess)
}

// availableModel is a single model entry in the available-models response.
type availableModel struct {
	Name string `json:"name"`
	Type string `json:"type"`
}

// availableModelsResponse is the JSON body returned by AvailableModels.
type availableModelsResponse struct {
	Models []availableModel `json:"models"`
}

// AvailableModels handles GET /api/v1/me/available-models.
// It returns the list of models accessible to the current key's scope,
// respecting the caller's key scope.
// Any authenticated key may call this endpoint — no additional role is required.
//
// @Summary      List models available to the authenticated key
// @Description  Returns models accessible to the authenticated key.
// @Tags         auth
// @Produce      json
// @Success      200  {object}  availableModelsResponse
// @Failure      401  {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /me/available-models [get]
func (h *Handler) AvailableModels(c fiber.Ctx) error {
	keyInfo := auth.KeyInfoFromCtx(c)
	if keyInfo == nil {
		return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "missing authentication")
	}

	ctx := c.Context()
	dbModels, err := h.DB.ListActiveModels(ctx)
	if err != nil {
		h.Log.ErrorContext(ctx, "available models: list active", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to list models")
	}

	models := make([]availableModel, 0, len(dbModels))
	for _, m := range dbModels {
		hasRoute := false
		steps, stepErr := h.DB.ListModelRouteSteps(ctx, m.ID, true)
		if stepErr != nil {
			h.Log.ErrorContext(ctx, "available models: list route steps",
				slog.String("model_id", m.ID),
				slog.String("error", stepErr.Error()),
			)
		} else if len(steps) > 0 {
			hasRoute = true
		}
		if !hasRoute {
			deps, depErr := h.DB.ListActiveDeployments(ctx, m.ID)
			if depErr != nil {
				h.Log.ErrorContext(ctx, "available models: list deployments",
					slog.String("model_id", m.ID),
					slog.String("error", depErr.Error()),
				)
				continue
			}
			if len(deps) == 0 {
				continue
			}
		}
		modelType := m.ModelType
		if modelType == "" {
			modelType = "chat"
		}
		models = append(models, availableModel{Name: m.Name, Type: modelType})
	}

	return c.JSON(availableModelsResponse{Models: models})
}

// changeOwnPasswordRequest is the JSON body for POST /api/v1/me/password.
type changeOwnPasswordRequest struct {
	CurrentPassword string `json:"current_password"`
	NewPassword     string `json:"new_password"`
}

// ChangeOwnPassword handles POST /api/v1/me/password. It verifies the caller's
// current password, then replaces it with the new one and invalidates all other
// active sessions for the user. The current session remains valid.
// Only password-based accounts may use this endpoint.
//
// @Summary      Change own password
// @Description  Verifies the current password and sets a new one. Revokes all other active sessions. Requires a user-scoped session key.
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        body  body      changeOwnPasswordRequest  true  "Password change request"
// @Success      200   "OK"
// @Failure      400   {object}  swaggerErrorResponse
// @Failure      401   {object}  swaggerErrorResponse
// @Failure      500   {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /me/password [post]
func (h *Handler) ChangeOwnPassword(c fiber.Ctx) error {
	keyInfo := auth.KeyInfoFromCtx(c)
	if keyInfo == nil {
		return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "missing authentication")
	}
	if keyInfo.UserID == "" {
		return apierror.Send(c, fiber.StatusBadRequest, "bad_request", "this endpoint requires a user-scoped key")
	}

	var req changeOwnPasswordRequest
	if err := c.Bind().JSON(&req); err != nil {
		return apierror.BadRequest(c, "invalid request body")
	}
	if req.CurrentPassword == "" {
		return apierror.BadRequest(c, "current_password is required")
	}
	if len(req.NewPassword) < 8 {
		return apierror.BadRequest(c, "new_password must be at least 8 characters")
	}
	if len(req.NewPassword) > 72 {
		return apierror.BadRequest(c, "new_password must be at most 72 bytes")
	}

	ctx := c.Context()

	authProvider, currentHash, err := h.DB.GetUserPasswordHashByID(ctx, keyInfo.UserID)
	if err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "user not found")
		}
		if errors.Is(err, db.ErrNoPassword) {
			// Burn bcrypt time to prevent timing-based account-type enumeration.
			// The error from CompareHashAndPassword is intentionally ignored here.
			_ = bcrypt.CompareHashAndPassword(dummyHash, []byte(req.CurrentPassword))
			return apierror.BadRequest(c, "password change not available for this account")
		}
		h.Log.ErrorContext(ctx, "change own password: get password hash", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to verify current password")
	}

	// External-auth accounts never have a local password hash.
	// The ErrNoPassword sentinel above covers the NULL hash case; this guard
	// catches rows where auth_provider is set to a non-local value but the hash
	// column is somehow populated — defensive, belt-and-suspenders.
	if authProvider != "local" {
		// Burn bcrypt time to prevent timing-based account-type enumeration.
		// The error from CompareHashAndPassword is intentionally ignored here.
		_ = bcrypt.CompareHashAndPassword(dummyHash, []byte(req.CurrentPassword))
		return apierror.BadRequest(c, "password change not available for this account")
	}

	if err := bcrypt.CompareHashAndPassword([]byte(currentHash), []byte(req.CurrentPassword)); err != nil {
		return apierror.Send(c, fiber.StatusBadRequest, "bad_request", "current password is incorrect")
	}

	newHash, err := bcrypt.GenerateFromPassword([]byte(req.NewPassword), bcrypt.DefaultCost)
	if err != nil {
		if errors.Is(err, bcrypt.ErrPasswordTooLong) {
			return apierror.BadRequest(c, "new_password must be at most 72 bytes")
		}
		h.Log.ErrorContext(ctx, "change own password: bcrypt", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to hash new password")
	}
	newHashStr := string(newHash)

	// Update the password hash and revoke all other sessions atomically.
	// If either operation fails the transaction rolls back, leaving the DB
	// consistent: neither the password nor the session list is partially updated.
	if err := h.DB.ChangePasswordAndRevokeOtherSessions(ctx, keyInfo.UserID, newHashStr, keyInfo.ID); err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "user not found")
		}
		h.Log.ErrorContext(ctx, "change own password: update password and revoke sessions", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to update password")
	}

	// Evict all other session entries for this user from the in-memory key cache
	// so that revoked sessions are rejected immediately without waiting for the
	// next periodic cache refresh.
	// Collect hashes first — Range holds a read lock, so Delete (write lock)
	// must not be called inside the Range closure.
	var toEvict []string
	currentKeyID := keyInfo.ID
	h.KeyCache.Range(func(keyHash string, ki auth.KeyInfo) bool {
		if ki.UserID == keyInfo.UserID && ki.ID != currentKeyID {
			toEvict = append(toEvict, keyHash)
		}
		return true
	})
	for _, keyHash := range toEvict {
		h.KeyCache.Delete(keyHash)
	}

	if h.AuditLogger != nil {
		h.AuditLogger.Log(audit.Event{
			Timestamp:    time.Now().UTC(),
			ActorID:      keyInfo.UserID,
			ActorType:    "user",
			ActorKeyID:   keyInfo.ID,
			Action:       "password_change",
			ResourceType: "user",
			ResourceID:   keyInfo.UserID,
			IPAddress:    c.IP(),
			StatusCode:   fiber.StatusOK,
			RequestID:    apierror.RequestIDFromCtx(c),
		})
	}

	return c.SendStatus(fiber.StatusOK)
}

// Me returns the authenticated user's profile.
//
// @Summary      Get authenticated user profile
// @Description  Returns the profile of the user associated with the current session key.
// @Tags         auth
// @Produce      json
// @Success      200  {object}  meResponse
// @Failure      400  {object}  swaggerErrorResponse
// @Failure      401  {object}  swaggerErrorResponse
// @Failure      404  {object}  swaggerErrorResponse
// @Failure      500  {object}  swaggerErrorResponse
// @Security     BearerAuth
// @Router       /me [get]
func (h *Handler) Me(c fiber.Ctx) error {
	keyInfo := auth.KeyInfoFromCtx(c)
	if keyInfo == nil {
		return apierror.Send(c, fiber.StatusUnauthorized, "unauthorized", "missing authentication")
	}
	if keyInfo.UserID == "" {
		return apierror.Send(c, fiber.StatusBadRequest, "bad_request", "this endpoint requires a user-scoped key")
	}
	user, err := h.DB.GetUser(c.Context(), keyInfo.UserID)
	if err != nil {
		if errors.Is(err, db.ErrNotFound) {
			return apierror.NotFound(c, "user not found")
		}
		h.Log.ErrorContext(c.Context(), "me: get user", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to retrieve user profile")
	}
	me, err := h.buildMeResponse(c.Context(), user, keyInfo.Role)
	if err != nil {
		h.Log.ErrorContext(c.Context(), "me: build response", slog.String("error", err.Error()))
		return apierror.InternalError(c, "failed to retrieve user profile")
	}
	_ = h.DB.TouchSessionLastUsed(c.Context(), keyInfo.ID, 5*time.Minute)
	return c.JSON(me)
}